Object Validation.

+++ database/migrations/2020_12_24_111528_create_purchases_table.php
@@ -30,7 +30,7 @@ class CreatePurchasesTable extends Migration
             $table->date('date_of_purchase');
             $table->date('start_date_of_contract')->nullable();
             $table->date('end_date_of_contract')->nullable();            
-            $table->integer('percentage_of_payment');
+            $table->float('percentage_of_payment');
             $table->unsignedInteger('cost');
             $table->unsignedInteger('amount_paid');            
             $table->date('next_payment_period')->nullable();
